tuning inline; report # sequences processed
When optimization is enabled, gcc and icc inline all static functions that are
called only once. Debuggers and profilers may be confused by this behavior.
Option "-fno-inline-functions" disables the feature in icc and
"-fno-inline-functions-called-once" disables in gcc. The latter option is not
implemented in icc, which will give a warning.